db연동, 모듈화
