+++ requirements.txt
@@ -1,0 +1,1 @@
+PyMySQL==0.9.3
